Split ffmpeg
That will simplify transfer to export parameters later
